rev, zip
