+++ src/main/scala/magic_rdds/package.scala
@@ -1,6 +1,8 @@
 
 package object magic_rdds
-  extends size
+  extends size  /** must come first because extends [[org.hammerlab.magic.rdd.cache.RDDCache]] class */
+    with rev
     with scan
     with sliding
     with sort
+    with zip
