+++ src/main/scala/magic_rdds/rev.scala
@@ -1,0 +1,4 @@
+package magic_rdds
+
+trait rev extends org.hammerlab.magic.rdd.rev
+object rev extends rev
